feat: run Cypress test against pre-compiled HTML
During the build html files are generated based on ascii doc files. When we run the Cypress test in a Maven build we should test against these html files for the challenge text etc. This way we make sure the asciidoc plugin in Maven works.

The Cypress test running outside Maven still use the on-the-fly asciidoc generator. This way one can still directly see the changes in the challenge text when running the Cypress tests directly.